refactor(data-fetcher): introduce a dataset class to give more control
The dataset instances are now responsible for fetching the data. To retain
previous behaviour, a preliminary HEAD request might be done by the browser
to check the advertised mime type in case the file/data format is unknown.
