+++ libs/util/data-fetcher/src/parsers/excel.ts
@@ -1,12 +1,13 @@
-import type { Feature } from 'geojson'
 import { jsonToGeojsonFeature } from './json'
+import { DataItem } from '../lib/model'
+import { BaseDataset } from './base'
 
 /**
  * This will read the first sheet of the excel workbook and expect the first
  * line to contain the properties names
  * @param buffer
  */
-export function parseExcel(buffer: ArrayBuffer): Promise<Feature[]> {
+export function parseExcel(buffer: ArrayBuffer): Promise<DataItem[]> {
   return import('xlsx').then(({ read, utils }) => {
     const workbook = read(buffer)
     const sheet = workbook.Sheets[workbook.SheetNames[0]]
@@ -18,4 +19,10 @@ export function parseExcel(buffer: ArrayBuffer): Promise<Feature[]> {
 
     return json.map(jsonToGeojsonFeature)
   })
+}
+
+export class ExcelDataset extends BaseDataset {
+  readAll(): Promise<DataItem[]> {
+    return this.fetchAsBuffer().then(parseExcel)
+  }
 }
