+++ libs/util/data-fetcher/src/parsers/json.ts
@@ -1,10 +1,11 @@
-import type { Feature } from 'geojson'
+import { DataItem } from '../lib/model'
+import { BaseDataset } from './base'
 
 /**
  * This parser only supports arrays of simple flat objects with properties
  * @param text
  */
-export function parseJson(text: string): Feature[] {
+export function parseJson(text: string): DataItem[] {
   const parsed = JSON.parse(text)
   if (!Array.isArray(parsed)) {
     throw new Error('Could not parse JSON, expected an array at root level')
@@ -12,7 +13,7 @@ export function parseJson(text: string): Feature[] {
   return (parsed as any[]).map(jsonToGeojsonFeature)
 }
 
-export function jsonToGeojsonFeature(object: { [key: string]: any }): Feature {
+export function jsonToGeojsonFeature(object: { [key: string]: any }): DataItem {
   const { id, properties } = Object.keys(object)
     .map((property) => (property ? property : 'unknown')) //prevent empty strings
     .reduce(
@@ -33,5 +34,11 @@ export function jsonToGeojsonFeature(object: { [key: string]: any }): Feature {
     geometry: null,
     properties,
     ...(id !== undefined && { id }),
+  }
+}
+
+export class JsonDataset extends BaseDataset {
+  readAll(): Promise<DataItem[]> {
+    return this.fetchAsText().then(parseJson)
   }
 }
